% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{abc}
\alias{abc}
\title{return a single letter from alphabet}
\usage{
abc(x, capitals = FALSE)
}
\arguments{
\item{x}{integer. Number of letter in the alphabet}

\item{capitals}{logical. Return uppercase letters?}
}
\value{
character.
}
\description{
return a single letter from alphabet
}
\examples{
    abc(1)
    abc(x=7, capitals = TRUE)
    abc(x=c(1:3))
}
\author{
Simon Frey
}
