% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{addSlash}
\alias{addSlash}
\title{Add a slash to an existing character string}
\usage{
addSlash(x)
}
\arguments{
\item{x}{character string}
}
\value{
character string
}
\description{
Adds a slash (/) to x, if x does not already end with one.
}
\examples{
addSlash("C:/TEMP")
addSlash("C:/TEMP/")
}
\seealso{
\code{\link{removeSlash}} for removing a trailing slash; \code{\link{changeSlash}} for substituting a slah with a double backslash
}
\author{
Simon Frey
}
