% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_COSEROreg_par.r
\name{adist_dels}
\alias{adist_dels}
\title{Approximate Distance allowing only for deletions}
\usage{
adist_dels(x, y, dels = NULL, returnstrings = FALSE, warn = FALSE)
}
\arguments{
\item{x}{string or vector of strings that the function will try to transform into y}

\item{y}{string that the function uses to transform x into}

\item{dels}{NULL or integer. If NULL (the default) set no limitations on how many deletions are allowed. If dels is an integer, allow only that many deletions as dels is set.}

\item{returnstrings}{logical. Should the strings be returned (TRUE) or the positions within x.}

\item{warn}{logical. Activate or deactivate warnings.}
}
\value{
a string (vector) or a numerical (vector)
}
\description{
Calculate the approximate distance of two strings returning the matches, where deletions where necessary only.
}
\details{
This function calls \code{adist} to calculate the approximate distance and the operations (insertions, substitutions and deletions) necessary to transform string x in string y. It then evaluates only the cases, where deletions where made only. The user may set a maximum of allowed deletions. If the transformation fails it throws a warning (if warn == TRUE) and return NA.
}
\examples{
   y <- c("PCOR", "TCOR", "TMMON", "INTMAX", "ETVEGCOR", "DAYSDRY", "DAYSWET", "ETSYSCOR")
   x <- "PCor1"
   
   adist_dels(x,y, dels = 1, returnstrings = TRUE)
   
   x <- "PCORXXX"
   
   adist_dels(x,y, dels = 1)
   adist_dels(x,y, dels = NULL)
}
\author{
Simon Frey
}
