% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.snow.state.r
\name{analyse.snow.state}
\alias{analyse.snow.state}
\title{Analyse snow state}
\usage{
analyse.snow.state(state, NB, IZ, read.file = FALSE, show.plot = FALSE, ...)
}
\arguments{
\item{state}{output of \code{\link{read.state}} or, if read.file == TRUE, a string pointing to a state file.}

\item{NB}{numeric. Number of subbasin that will be analysed.}

\item{IZ}{numeric. Number of cell in the subbasin NB that will be analysed.}

\item{read.file}{logical. If TRUE a state file will be read in. In this case, state has to be a string pointing to a state file.}

\item{show.plot}{logical. Should the result be plotted?}

\item{...}{additional arguments passed on to \code{\link{dev.new.file}}}
}
\value{
a list containing the state of the snow cover on grid scale (1) and subgrid scale (2)
}
\description{
Analyse the state of the snow cover in a state file
}
\examples{
    x <- read.state("C:/TEMP/statevar.dmp")
    analyse.snow.state(state = x,NB=1,IZ=1)
    
    analyse.snow.state("C:/TEMP/statevar.dmp",NB=1,IZ=1,read.file=TRUE)
    
}
\seealso{
\code{\link{read.state}}
}
\author{
Simon Frey
}
