% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.daily.columns.r
\name{apply.daily.columns}
\alias{apply.daily.columns}
\title{Columnwise apply a function to receive hourly or daily aggregated values}
\usage{
apply.daily.columns(x, FUN, agg = "day", PB = TRUE, ncores = 0, tz = NULL, ...)
}
\arguments{
\item{x}{an xts object}

\item{FUN}{an R function}

\item{agg}{character string to specify whether to receive monthly, weekly, daily or hourly values, respectively. See details.}

\item{PB}{logical. Should a progressbar be shown? See details.}

\item{ncores}{integer. Number of CPU cores used. See details.}

\item{tz}{character specifying the time zone or NULL (the standard). If the latter, the time zone of x is used.}

\item{...}{additional arguments to FUN}
}
\value{
An xts object containing hourly, daily, ... values
}
\description{
Apply a specified function to each column of an xts object creating hourly, daily or monthly values
}
\details{
A simple mechanism to use \code{\link{apply.daily}}, \code{\link{apply.hourly}},  \code{\link{apply.weekly}} or \code{\link{apply.monthly}} to each column of an xts object.

    Since November 2025 apply.daily.columns uses multicore parallel computing. By default it uses all but one cores on the system. By setting ncores, the user may overrule this.
    
    In previous versions, PB indicated whether a txt or winProgressbar should indicate the progress of the calculations. Since parallel computing is now supported PB now indicates whether a progressbar should be displayed (using \code{\link{tkProgressBar}}). Default is TRUE. Still "n" suppresses the progressbar. The old parameters ("t", "txt", "w" "win") are kept for compability reasons.
}
\examples{
 
    # load precipitation input
    data(precipitation)
    head(x)
    
    # aggregate to daily values
    aday <- apply.daily.columns(x, FUN = sum, agg = 'day', PB = TRUE)
    head(aday)
}
\seealso{
\code{\link{apply.daily}}

\code{\link{apply.hourly}}

\code{\link{apply.monthly}}

\code{\link{apply.weekly}}
}
\author{
Simon Frey
}
