% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{changeSlash}
\alias{changeSlash}
\title{Substitute slashes by double backslashes}
\usage{
changeSlash(x)
}
\arguments{
\item{x}{character string}
}
\value{
charater string with double backslashes instead of slashes
}
\description{
Substitute slashes by double backslashes in a character string.
}
\examples{
changeSlash("C:/Temp")
changeSlash("C:/firstlevel/secondlevel")
}
\seealso{
See \code{\link{addSlash}} for adding a trailing slash and \code{\link{removeSlash}} for removing a trailing slash from a character string.
}
\author{
Simon Frey
}
