% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{clear.duplicated}
\alias{clear.duplicated}
\title{Remove duplicated entries}
\usage{
clear.duplicated(x)
}
\arguments{
\item{x}{xts object}
}
\value{
xts object
}
\description{
Searches duplicated timestamps in an xts object and removes them. The new value is the mean of all values with the same timestamp.
}
\author{
Simon Frey
}
