% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{colMax}
\alias{colMax}
\title{Find the max value of a (set of) column(s)}
\usage{
colMax(colData, na.rm = TRUE)
}
\arguments{
\item{colData}{numerical matrix}

\item{na.rm}{logical. Should NA values be removed?}
}
\value{
numerical vector
}
\description{
Find the max value of a (set of) column(s)
}
\examples{
data("runoff")
colMax(runoff)
}
\seealso{
\code{\link{colMin}}
}
\author{
Simon Frey
}
