% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{colMin}
\alias{colMin}
\title{Find the min value of a (set of) column(s)}
\usage{
colMin(x, na.rm = TRUE)
}
\arguments{
\item{x}{numerical matrix}

\item{na.rm}{logical. Should NA values be removed?}
}
\value{
numerical vector
}
\description{
Find the min value of a (set of) column(s)
}
\examples{
    data("runoff")
    colMin(runoff)
}
\seealso{
\code{\link{colMax}}

\code{\link{rowMin}}

\code{\link{rowMax}}

\code{\link{rowMedian}}
}
\author{
Simon Frey
}
