% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcolstats.r
\name{colStats}
\alias{colStats}
\title{Calculate some simple statistic estimators on cols in a matrix or data.frame}
\usage{
colStats(x, fun, na.rm = TRUE, q = 0.1)
}
\arguments{
\item{x}{numerical matrix or data.frame.}

\item{fun}{statistical function that will be calculated. See details.}

\item{na.rm}{logical. Should NA values be removed?}

\item{q}{numerical. Quantile that will be estimated. Used only if fun == quantile}
}
\value{
numerical vector
}
\description{
Calculate some simple statistics like median, min, max or quantiles on columns in a matrix or data.frame object
}
\details{
This function uses predefined statistical function that can be used:
\itemize{
\item min (determine the minimum value)
\item max (determine the maximum value)
\item mean (estimate the mean value)
\item median (estimate the median value)
\item quantile (estimate a certain quantile value. Note that values between 0 and 1 are accepted)
}
}
\examples{
    # load runoff data
    data(runoff)
    
    colStats(runoff, fun = "mean")
    colStats(runoff, fun = "max")
    colStats(runoff, fun = "quantile", q = 0.33)
}
\seealso{
\code{\link{colStats}}

\code{\link{colMin}}

\code{\link{rowMin}}

\code{\link{rowMax}}

\code{\link{colMax}}
}
\author{
Simon Frey
}
