% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{dev.new.file}
\alias{dev.new.file}
\title{Open a new device for plotting}
\usage{
dev.new.file(
  device = "dev",
  return.device = TRUE,
  filename = NULL,
  width = 480,
  height = 480,
  units = "in",
  res = 300,
  ...
)
}
\arguments{
\item{device}{any string of "dev","png","pdf", or "svg"}

\item{return.device}{logical. should the type of device be returned?}

\item{filename}{string. Where should the plot be saved? If no extension is given, it will be added according to the type of the device.}

\item{width}{numeric. Width of the graphics device in units.}

\item{height}{numeric. Height of the graphics device in units.}

\item{units}{string. The units in which height and width are given. Can be px (pixels), in (inches), cm or mm. By default it is "in".}

\item{res}{numeric.     The nominal resolution in ppi which will be recorded in the bitmap file, if a positive integer. Also used for units other than the default.}

\item{...}{additional arguments passed on to the dev function (e.g. \link{dev.new}, or \link{png}).}
}
\description{
Open a new device for plotting and allow some settings
}
\details{
Note that the default units in inches. This can be changed by defining the 'units' string. However, it only has an effect when the image is printed as png.
}
\examples{
    d <- dev.new.file(device = "png",filename = "C:/TEMP/test")
    plot(1)
    if(d \%in\% c("png","pdf","svg")) dev.off()

}
\author{
Simon Frey
}
