% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.r
\name{distrib}
\alias{distrib}
\title{compute snow distribution accoding to nvar}
\usage{
distrib(SWE, nvar)
}
\arguments{
\item{SWE}{amount of (new) snow to be dirstibuted over a grid cell}

\item{nvar}{parameter from COSERO}
}
\value{
A vector of length 5 with the snow distribution
}
\description{
compute snow distribution in COSERO according to the parameter nvar
}
\details{
formula : SWE(IKL) = exp(m + s * SWE)
    
    m = mean(SWE)
    s = sd(SWE)
    
    nvar = variance of SWE
    
    Only apply if 5 snow classes are used!
}
\examples{
    distrib(10,5)
    distrib(10,0)
}
\seealso{
\link{read_COSERO_par} for reading a COSERO parameter file
}
\author{
Simon Frey
}
