% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dy.cosero.r
\name{dy.cosero}
\alias{dy.cosero}
\title{Plot results regarding runoff from COSERO using dygraphs}
\usage{
dy.cosero(
  qoutput = NULL,
  read.data = TRUE,
  prec = FALSE,
  comp = FALSE,
  eta = FALSE,
  storage = FALSE,
  snowmelt = FALSE,
  group = NULL,
  area = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{qoutput}{Either a path to the file to be read or the result from \link{readCosero}}

\item{read.data}{logical. If TRUE, qoutput will be read in using \link{readCosero}}

\item{prec}{logical or character string. Should precipitation data be read in? Maybe a path to the corresponding file.}

\item{comp}{logical or character string. Should data about the composition of runoff be read in? Maybe a path to the corresponding file.}

\item{eta}{logical or character string. Should data about evapotranspiration be read in? Maybe a path to the corresponding file.}

\item{storage}{logical or character string. Should storage data be read in? Maybe a path to the corresponding file.}

\item{snowmelt}{logical or character string. Should snow data be read in? Maybe a path to the corresponding file.}

\item{group}{character string or NULL. Used to group dygraph plots. See \link{dygraph}}

\item{height}{numeric or NULL. Height in pixels (optional, defaults to automatic sizing)}
}
\value{
Nothing is returned to R. Only the dygraphs are plotted.
}
\description{
Plot the results regarding runoff from COSERO using dygraphs. The output may be grouped together with other results.
}
\details{
If read,data == FALSE the results from \code{\link{readCosero}} must be passed on to \code{dy.cosero}.
    In this case the parameters \code{prec},\code{comp},\code{storage},\code{snowmelt}, and \code{eta} may be logical
    
    Note that at this time, this function only works for basins with only one subbasin!
}
\author{
Simon Frey
}
