% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_par.r
\name{edit_par}
\alias{edit_par}
\title{Edit a parameter in a COSERO parameter file}
\usage{
edit_par(file, val, param)
}
\arguments{
\item{file}{A previously loaded COSERO parameter file}

\item{val}{Value(s) to be edited. May be a single value (character or numeric) or a vector of values}

\item{param}{Name of the parameter that will be edited.}
}
\value{
A list of the same structure as \code{file}
}
\description{
Edit a parameter in a previouly loaded COSERO parameter file. Either a single value or a vector is accepted.
}
\details{
If val is of length=1, it will be repeated using \code{\link{rep}} to the length of parameter values in the parameter file. Use \code{\link{read_COSERO_par}} for reading a COSERO parameter file.
}
\seealso{
\link{read_COSERO_par} for reading a COSERO parameter file

\link{write_COSERO_par} for writing a COSERO parameter file
}
\author{
Simon Frey
}
