% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothxts.r
\name{embed.fullextent}
\alias{embed.fullextent}
\title{Modified version of embed}
\usage{
embed.fullextent(x, timesteps = timesteps)
}
\arguments{
\item{x}{a numeric vector, matrix or time series}

\item{timesteps}{integer. Number of timestpes used for smoothing.}
}
\value{
A matrix containing the embedded time series x
}
\description{
Modified version of \code{\link{embed}}.
}
\details{
By adding the last timesteps-1 entries at the end of the result of embed, the length of the output is equal to the length of x. Otherwise the length of the output would be length(x)-(timesteps-1)
}
\author{
Simon Frey
}
