% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{fill.missing}
\alias{fill.missing}
\title{Fill missing values in an xts object}
\usage{
fill.missing(x, start = NULL, end = NULL, steps = "hour")
}
\arguments{
\item{x}{an xts object}

\item{start}{NULL or an POSIXct object. If NULL, the first index of x will be used}

\item{end}{NULL or an POSIXct object. If NULL, the last index of x will be used}

\item{steps}{increment of the sequence. See ‘Details’.}
}
\value{
an xts object
}
\description{
Fill missing entries in an xts object using \code{\link{nalocf}}
}
\details{
steps can be a character string, containing one of "sec", "min", "hour", "day", "DSTday", "week", "month", "quarter" or "year". Default is "hour"
}
\seealso{
\code{\link{missing.steps}}
}
\author{
Simon Frey
}
