% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.parameter.r
\name{get.parameter}
\alias{get.parameter}
\title{Extract a parameter from a COSERO parameter file}
\usage{
get.parameter(x, parafile, dims = NULL)
}
\arguments{
\item{x}{character. The name of the parameter that is extracted}

\item{parafile}{character. The parameter file. May be a path to a file or a object retrieved by \code{\link{read_COSERO_par}}}

\item{dims}{a matrix giving the dimensions of the parafile or NULL. If the latter, the dimensions of parafile will be determined.}
}
\value{
A matrix consisting of NB, IZ, NZ and the parameter is returned
}
\description{
Extract a parameter from a COSERO parameter file.
}
\examples{
    ### do not run
    get.parameter("CTMAX",par)
}
\seealso{
\code{\link{read_COSERO_par}}
}
\author{
Simon Frey
}
