% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{is.leapyear}
\alias{is.leapyear}
\title{Check for leap years}
\usage{
is.leapyear(year)
}
\arguments{
\item{year}{numeric}
}
\value{
logical. TRUE if year is a leap year
}
\description{
Check for leap years
}
\examples{
is.leapyear(2000)
is.leapyear(2001)
is.leapyear(c(1970:2100))
}
\seealso{
\url{http://en.wikipedia.org/wiki/Leap_year}
}
\author{
Simon Frey
}
