% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.unique.r
\name{is.unique}
\alias{is.unique}
\title{Check for unique values}
\usage{
is.unique(x, ...)
}
\arguments{
\item{x}{vector to be analysed.}

\item{...}{additional parameters passed to \code{\link{count}}}
}
\value{
a boolean vector
}
\description{
Check whether a variable within a vector is unique
}
\details{
This function uses the \code{\link{count}} function from the plyr package and returns all entries with a frquency greater than one. Works with character and numeric vectors
}
\seealso{
\code{\link{unique}}
}
\author{
Simon Frey
}
