% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{m3s2mm}
\alias{m3s2mm}
\title{Convert m^3/s in mm}
\usage{
m3s2mm(input, area, dt)
}
\arguments{
\item{input}{numeric. May be a vector, matrix, timeseries, etc}

\item{area}{numeric. Area in m^2}

\item{dt}{numeric. time step}
}
\value{
numeric of the same type as input
}
\description{
Convert m^3/s in mm
}
\examples{
data("Nile")
m3s2mm(input = Nile, area = 6852000000, dt = 86400 * 365)
}
\author{
Simon Frey
}
