% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{missing.steps}
\alias{missing.steps}
\title{Find missing time steps in an univariate time series}
\usage{
missing.steps(x)
}
\arguments{
\item{x}{an xts object}
}
\value{
a numerical vector giving the differences of the time steps of the xts object

a numeric vector
}
\description{
Irregularities are found by \code{\link{diff.POSIXt}}. A majority vote using \code{\link{which.max}} from the package \code{table} determines the most common time step. 
Any differing timesteps are interpreted as missing. A vector containing the locations of those missing time steps is returned.
Note that only the location but not how many time steps per gap are missing is returned.
}
\seealso{
\code{\link{fill.missing}}
}
\author{
Simon Frey
}
