% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{mm2m3s}
\alias{mm2m3s}
\title{Convert mm to m^3/s}
\usage{
mm2m3s(input, area, dt)
}
\arguments{
\item{input}{numeric. May be a vector, matrix, timeseries, etc}

\item{area}{numeric. Area in m^2}

\item{dt}{numeric. time step}
}
\value{
numeric of the same type as input
}
\description{
Convert mm to m^3/s
}
\examples{
data("runoff")
mm2m3s(input = as.numeric(input[,1]), area = 20000000, dt = 3600)
mm2m3s(input = runoff, area = c(rep(20000000,3),rep(50000000,3)), dt = 3600)
}
\author{
Simon Frey
}
