% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{nalocf}
\alias{nalocf}
\title{Faster version of last observation carried forward}
\usage{
nalocf(x, loop = FALSE, ...)
}
\arguments{
\item{x}{an xts object}

\item{loop}{logical. Should the operation be performed as a loop (slower but maybe better for RAM management)}

\item{...}{further arguments, e.g. na.rm; see \code{\link{na.locf}}}
}
\value{
Returns an xts object
}
\description{
This is basically a wrapper around \code{\link{na.locf}} from the xts package. It avoids the column-by-column copying. Usefull if there are many columns in the xts object.
}
\examples{
data(runoff)
# add some NAs
runoff[2,] <- NA
index(runoff)[2] <- NA

a <- nalocf(runoff)
b <- na.locf(runoff)
identical(a,b)
}
\author{
Simon Frey
}
