% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdiff.r
\name{peakdiff}
\alias{peakdiff}
\title{Calculate the peak difference of two hydrographs}
\usage{
peakdiff(sim, obs, timesteps, nop = 1, FUN = me, ...)
}
\arguments{
\item{sim}{xts object. The simulated hydrograph.}

\item{obs}{xts object. The observed hydrograph.}

\item{timesteps}{numeric. How many timesteps should be considered (equal distributed around the observed peak)?}

\item{nop}{numeric. How many peaks should be analysed?}

\item{FUN}{function. Objective function to be calculated whitin the region of the peak(s).}

\item{...}{additional arguments passed to FUN}
}
\value{
a vector with the results of the objective function
}
\description{
Calculate the peak difference of two hydrographs.
}
\details{
FUN should be a function that can operate with sim and obs values and returns one single value.
}
\author{
Simon Frey
}
