% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better.xts.r
\name{plot.xxts}
\alias{plot.xxts}
\title{Advanced plotting function to plot xts objects}
\usage{
\method{plot}{xxts}(x, y = NULL, type = "l", auto.grid = TRUE,
  major.ticks = "auto", minor.ticks = TRUE, major.format = TRUE,
  bar.col = "grey", candle.col = "white", ann = TRUE, axes = TRUE,
  lines.col = "black", xlabels = TRUE, ...)
}
\arguments{
\item{x}{an xts object}

\item{y}{an xts object or NULL}

\item{type}{type of plot to produce}

\item{auto.grid}{should grid lines be drawn}

\item{major.ticks}{should major tickmarks be drawn and labeled}

\item{minor.ticks}{should minor tickmarks be drawn}

\item{major.format}{passed along to axTicksByTime.}

\item{bar.col}{the color of the bars when type is ‘bars’ or ‘candles’}

\item{candle.col}{the color of the candles when type is ‘candles’}

\item{ann}{passed ‘par’ graphical parameter}

\item{axes}{passed ‘par’ graphical parameter}

\item{lines.col}{the color of the candles when type is ‘lines’}

\item{xlabels}{logical or a vector giving the labels (see details)}

\item{...}{additional graphical arguments}
}
\value{
Plots an xts object to the current device.
}
\description{
The function is identical to the plotting function in the xts package but has the advantage
that labeling of the x-axis can be turned off. In addition, it takes the mgp options from par instead of using
the standard ootions without the possibility to edit them.
    Also, the lines.color can be set.
}
\details{
Mainly used to draw time-series plots with sensible x-axis labels, it can also plot basic OHLC series using type='candles' or type='bars'.

   Better financial plots can be found in the quantmod package, though these are generally incompatible with standard R graphics tools.

   If xlabels is FALSE, no lables are drawn (but the ticks are). If it is TRUE, the lables determined by the xts function are drawn. If it is a vector, this vector is used for labeling. Will be recycled.
}
\examples{
data(sample_matrix)
plot(sample_matrix)
plot.xxts(as.xts(sample_matrix))
plot.xxts(as.xts(sample_matrix), lines.col="red", xlabels = FALSE)
}
\seealso{
\code{\link{xts}}
}
\author{
Simon Frey
}
