% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{read.qobsqsim}
\alias{read.qobsqsim}
\title{read qobs_qsim.txt (a COSERO output file)}
\usage{
read.qobsqsim(x, ...)
}
\arguments{
\item{x}{character string pointing towards the file qobs_qsim.txt}

\item{...}{other arguments passed to \code{\link{read.table}}. Already defined are \code{header (TRUE)} and \code{colClasses}}
}
\value{
a list of length 2 with QOBS and QSIM values, respectively
}
\description{
Reads a qobs_qsim.txt file which is written by COSERO. Returns the recordings as list holding xts objects.
}
\details{
The file qobs_qsim.txt typically is located in the output folder of COSERO. However, by default, COSERO does NOT  write this file. To tell COSERO to write this, use OUTCONTROL = 1 or 2 in the COSERO defaults file.

    For reading Q_output.txt, which is always be written by COSERO, see \code{\link{readCosero}}.
}
\examples{
fpath <- system.file("extdata","qobs_qsim.txt", package = "TigR")
out <- read.qobsqsim(fpath)
summary(out)
}
\seealso{
\code{\link{readCosero}}
}
\author{
Simon Frey
}
