% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{read.xts}
\alias{read.xts}
\title{read a textfile and save it as xts-object}
\usage{
read.xts(x, datecolumns = c(1:5), format = "\%Y \%m \%d \%H \%M",
  header = TRUE, tz = "utc", skip = 0, ...)
}
\arguments{
\item{x}{character string. file to load}

\item{datecolumns}{integer vector. Giving the columns with date informations}

\item{format}{character string. Specify the format of the date}

\item{header}{logical. Does the data have a header?}

\item{tz}{character string. Time zone of the data}

\item{skip}{integer: the number of lines of the data file to skip before beginning to read data}

\item{...}{additional arguments from other methods passed to read.table}
}
\value{
an xts object.
}
\description{
Read a text file and directly save it as xts object.
}
\examples{
fpath <- system.file("extdata","qobs_qsim.txt", package = "TigR")
out <- read.xts(x = fpath)
}
\author{
Simon Frey
}
