% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCOSEROpar.r
\name{read_COSERO_par}
\alias{read_COSERO_par}
\title{Read a COSERO parameterfile}
\usage{
read_COSERO_par(x, NB = NULL)
}
\arguments{
\item{x}{character string. A COSERO parameter file}

\item{NB}{numerical. Return only the parameters of a certain subbasin. May be a single number or a numerical vector.}
}
\value{
A list containing three lists: Nr 1 contains the dimensions, Nr 2 contains the headers, and number 3 the values
}
\description{
Read a COSERO parameter file and return the values as list.
}
\author{
Simon Frey
}
