% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcosero.r
\name{readCosero}
\alias{readCosero}
\title{Reading COSERO output files}
\usage{
readCosero(qoutput = "./Q_output.txt", prec = FALSE, comp = FALSE,
  eta = FALSE, storage = FALSE, snowmelt = FALSE, area = NULL,
  timestep = NULL, ...)
}
\arguments{
\item{qoutput}{character string. Pointing towards the file "Q_output.txt"}

\item{prec}{logical. Should precipitation data be read in?}

\item{comp}{logical. Should data about the composition of runoff be read in?}

\item{eta}{logical. Should data about evapotranspiration be read in?}

\item{storage}{logical. Should storage data be read in?}

\item{snowmelt}{logical. Should snow data be read in?}

\item{area}{numeric. Area of the subbasin(s).}

\item{timestep}{numeric. Timestep of the data.}

\item{...}{Additional arguments passed to read.table from other methods.}
}
\description{
Read COSERO outputfiles and store them into a list.
}
\details{
Reading the file Q_output.txt.

   The function automatically searches for the file containing data about precipitation, runoff composition, eta, storage or snowmelt if the respective arguments are set TRUE. Note that those file are NOT written by COSERO by default. Use OUTCONTROL = 1 or 2 in the defaults file of COSERO to activate writing of them. They are given in mm. To compare them to runoff, be adviced that runoff should be transformed into mm, too. This, however, can only be done if both area is given. It is done via \code{\link{m3s2mm}}. The argument timestep my be provided. If NULL the function will try to gather it from the data.

   For reading qobs_qsim.txt see \code{\link{read.qobsqsim}}
}
\examples{
data(area)
fpath <- system.file("extdata","Q_output.txt", package = "ownfunc")
x <- readCosero(qoutput = fpath, area = area)
x <- readCosero(qoutput = fpath)
}
\seealso{
\code{\link{m3s2mm}}

\code{\link{read.qobsqsim}}
}
\author{
Simon Frey
}
