% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHZB.r
\name{readHZB}
\alias{readHZB}
\title{read a file provides by HZB}
\usage{
readHZB(
  file,
  name = TRUE,
  keyword = "Messstelle",
  format = "\%d.\%m.\%Y \%H:\%M",
  tz = "UTC",
  ...
)
}
\arguments{
\item{file}{file to read}

\item{name}{logical. Should the return value contain a name given by the keyword in the file?}

\item{keyword}{character. Keyword used for naming the xts object.}

\item{format}{format for the formatting date and time (see \code{\link{as.POSIXct}})}

\item{tz}{timezone}

\item{...}{agruments passed on to \code{\link{read.xts}}}
}
\value{
an xts object
}
\description{
read a file provided by HZB (e.g. from ehyd)
}
\examples{
 ### not run ###
 readHZB(x, TRUE)
   
}
\seealso{
\code{\link{read.table}}

\code{\link{readLines}}

\code{\link{read.xts}}
}
\author{
Simon Frey
}
