% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{removeSlash}
\alias{removeSlash}
\title{Remove a trailing slash from a character string}
\usage{
removeSlash(x)
}
\arguments{
\item{x}{character string}
}
\value{
character string without a trailing slash
}
\description{
Removes a trailing slash (/) from x
}
\examples{
removeSlash("C:/Temp/")
removeSlash("C:/Temp")
}
\seealso{
\code{\link{addSlash}} for adding a trailing slash
}
\author{
Simon Frey
}
