% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.dls.r
\name{rm.dls}
\alias{rm.dls}
\title{Remove daylight saving time steps in an extended time series}
\usage{
rm.dls(x)
}
\arguments{
\item{x}{an xts object.}
}
\value{
an xts object.
}
\description{
Remove the daylight saving time steps in an xts-object
}
\details{
The function essentially creates an time series sequence in utc with the same length as x. Then the index of x is replaced by this new time series.
}
\author{
Simon Frey
}
