% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.turbines.r
\name{rm_turbine}
\alias{rm_turbine}
\title{remove influence of turbine activity from a time series}
\usage{
rm_turbine(x, rise = 300, tolerance = 100, duration = 12, thresQ = NULL)
}
\arguments{
\item{x}{an xts object}

\item{rise}{numeric. Rise in percent that needs to be exceeded for detecting as turbine influence}

\item{tolerance}{numeric. tolerance in percent that the runoff level needs reach again}

\item{duration}{numeric. time steps that will be analyzed after an exceed event was detected}

\item{thresQ}{numeric or NULL. Threshold of runoff below which no detection will be performed.}
}
\value{
an xts object
}
\description{
analyse a xts object and try to remove the activity from turbines from it
}
\author{
Simon Frey
}
