% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{rna}
\alias{rna}
\title{Alternative to na.locf}
\usage{
rna(z)
}
\arguments{
\item{z}{data.frame or matrix}
}
\description{
honestly, can't remember why and what it does
}
\author{
Simon Frey
}
