% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{rowMax}
\alias{rowMax}
\title{Find the max value of a (set of) row(s)}
\usage{
rowMax(x, na.rm = TRUE)
}
\arguments{
\item{x}{numerical matrix}

\item{na.rm}{logical. Should NA values be removed?}
}
\value{
numerical vector
}
\description{
Find the max value of a (set of) row(s)
}
\examples{
data("runoff")
rowMax(runoff)
}
\seealso{
\code{\link{colMin}}

\code{\link{rowMin}}

\code{\link{colMax}}

\code{\link{rowMedian}}
}
\author{
Simon Frey
}
