% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{rowMedian}
\alias{rowMedian}
\title{Calculate the median of a (set of) row(s)}
\usage{
rowMedian(x, na.rm = TRUE)
}
\arguments{
\item{x}{numerical matrix}

\item{na.rm}{logical. Should NA values be removed?}
}
\value{
numerical vector
}
\description{
Calculate the median of a (set of) row(s)
}
\examples{
data("runoff")
rowMedian(runoff)
}
\seealso{
\code{\link{colMin}}

\code{\link{rowMin}}

\code{\link{rowMax}}

\code{\link{colMax}}
}
\author{
Simon Frey
}
