% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{rowMin}
\alias{rowMin}
\title{Find the min value of a (set of) row(s)}
\usage{
rowMin(x, na.rm = TRUE)
}
\arguments{
\item{x}{numerical matrix}

\item{na.rm}{logical. Should NA values be removed?}
}
\value{
numerical vector
}
\description{
Find the min value of a (set of) row(s)
}
\examples{
    data("runoff")
    rowMin(runoff)
}
\seealso{
\code{\link{colMin}}

\code{\link{colMax}}

\code{\link{rowMax}}

\code{\link{rowMedian}}
}
\author{
Simon Frey
}
