% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcolstats.r
\name{rowStats}
\alias{rowStats}
\title{Calculate some simple statistic estimators on rows in a matrix, data.frame, or xts object}
\usage{
rowStats(
  x,
  fun,
  na.rm = TRUE,
  q = 0.1,
  weights = NULL,
  run.parallel = FALSE,
  keep.plan = FALSE
)
}
\arguments{
\item{x}{numerical matrix or data.frame.}

\item{fun}{statistical function that will be calculated. See details.}

\item{na.rm}{logical. Should NA values be removed?}

\item{q}{numerical. Quantile that will be estimated. Used only if fun == quantile.}

\item{weights}{numerical vector. Used if fun == weighted.quantile only. See details.}

\item{run.parallel}{logical. Should the operation be run in parallel? If TRUE, \link{future_apply} with the plan "multisession" is used.}

\item{keep.plan}{logical. Should the plan "multisession" be kept or changed back to "sequential" by quitting the function?}
}
\value{
numerical vector, if x is an xts object, an xts object will be returned, too.
}
\description{
Calculate some simple statistics like median, min, max or quantiles on rows in a matrix, data.frame, or xts object
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function uses predefined statistical function that can be used:
* min (determine the minimum value)
* max (determine the maximum value)
* mean (estimate the mean value)
* median (estimate the median value)
* quantile (estimate a certain quantile value. Note that values between 0 and 1 are accepted)
* weighted.quantile (calculation of quantiles using individual weights. \link{Quantile} from \code{DescTools} is used.)
}\if{html}{\out{</div>}}
}
\examples{
    # load runoff data
    data(runoff)
    
    rowStats(runoff, fun = "mean")
    rowStats(runoff, fun = "max")
    rowStats(runoff, fun = "quantile", q = 0.33)
    rowStats(runoff, fun = "weighted.quantiles", q = 0.33, weights = c(1,1,1,4,1,1))
}
\seealso{
\code{\link{colStats}}

\code{\link{colMin}}

\code{\link{rowMin}}

\code{\link{rowMax}}

\code{\link{colMax}}
}
\author{
Simon Frey
}
