% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcosero.r
\name{runcosero}
\alias{runcosero}
\title{Run the hydrological model COSERO}
\usage{
runcosero(path, saveold = TRUE, batchfile = NA, use.statevar = TRUE,
  which.copy = "all", show.output = FALSE, ...)
}
\arguments{
\item{path}{Path to the executable of COSERO.}

\item{saveold}{logical. Save an old model run?}

\item{batchfile}{character string. Use a predefined batchfile for running COSERO or start COSERO with standard settings. In the latter case, set batchfile = NA.}

\item{use.statevar}{logical. Should COSERO read its statevariables form statevar.dmp within its input dir?}

\item{which.copy}{chracter string. Which files of the old model run should be saved. Accepts either a vector of filenames or "all".}

\item{show.output}{logical. Show the console output of COSERO in the console of R?}

\item{...}{Further arguments passed to \code{\link{shell}}}
}
\value{
Returns nothing to R but executes the model COSERO
}
\description{
Call the hydrological model COSERO via a shell command
}
\details{
runcosero calls a batchfile via the command \code{\link{shell}}. In this batchfile, COSERO is fed with the necessary user input.
The user can provide a batchfile (e.g. with specific commands to COSERO) or let runcosero write the batchfile itself.
In the first case, batchfile must not be NA but a path to the user defined batchfile. In the latter case, both the batchfile and the necessary commandsfile are deleted again. 

The user can decide if COSERO should read its stave variales from a statevar.dmp file or use standard state variables via the option use.statevar.

path must be set to a full path to the executable of COSERO including the exe itself. If the exe is not specified, "COSERO.exe" is assumed.
}
\examples{
runcosero(path = "C:/COSERO/COSERO.exe", saveold = TRUE, which.copy = "all", show.output = TRUE)
}
\references{
Frey, S., and Holzmann, H.: A conceptual, distributed snow redistribution model. Hydrol. Earth Syst. Sci., 19, 4517–4530, doi:10.5194/hess-19-4517-2015 2015
   
   Herrnegger, M., Nachtnebel, H.-P., and Haiden, T.: Evapotranspiration in high alpine catchments – an important part of the water balance!, Hydrol. Res., 43, 460–475, doi:10.2166/nh.2012.132, 2012.
   
   Kling, H., Stanzel, P., Fuchs, M., and Nachtnebel, H.-P.: Performance of the COSERO precipitation-runoff model under nonstationary conditions in basins with different climates, Hydrolog. Sci. J., 60, 1374–1393, doi:10.1080/02626667.2014.959956, 2015
}
\seealso{
For reading the output of COSERO see \code{\link{readcosero}} and \code{\link{read.qobsqsim}}
}
\author{
Simon Frey
}
