% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{scalevalues}
\alias{scalevalues}
\title{Normalize values between 0 and 1}
\usage{
scalevalues(x, ...)
}
\arguments{
\item{x}{numeric (vector, matrix or data.frame) that will be scaled}

\item{...}{further arguments passed to the functions min and max e.g. na.rm}
}
\value{
depending on x a normalized vector, matrix or data.frame is returned
}
\description{
Normalize values of a vector, matrix or data.frame between 0 and 1
}
\examples{
    ex <- c(3,5,10,2,9,20)
    scalevalues(ex)
}
\author{
Simon Frey
}
