% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_year.r
\name{set_year}
\alias{set_year}
\title{Alter the year in a POSIX-Object or Series}
\usage{
set_year(x, year = NULL)
}
\arguments{
\item{x}{a POSIX object or a vector of such objects}

\item{year}{character or character vector. The "new" year that the new POSIX object(s) will be hold}
}
\value{
A POSIX object or a vector of POSIX objects
}
\description{
Alter the year in a POSIX object or a vector of such objects
}
\details{
If year == NULL the year of \link{Sys.Date} will be used. Note, that, if year is a single character value, it is interpreted as the first value in case that x is a vector.
}
\examples{
    x <- as.POSIXct(c("2022-12-30","2022-12-31","2023-01-01","2023-01-02"))
    year <- "2023"
    set_year(x, year)
    
    #################################
    
    x <- as.POSIXct("2021-01-01")
    year <- "2023"
    set_year(x,year)
    
    #################################
     
    x <- as.POSIXct(c("2022-12-30","2022-12-31","2023-01-01","2023-01-02"))
    year <- c("2023","2023","2024","2024")
    set_year(x,year)
    
    #################################
}
\author{
Simon Frey
}
