% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{setcolumns}
\alias{setcolumns}
\title{Define a mfrow object out of a number of plots}
\usage{
setcolumns(x, maxrow = 4)
}
\arguments{
\item{x}{number of plots to be drawn}

\item{maxrow}{maximum number of plots in one row}
}
\value{
integer vector of length 2
}
\description{
returns an integer vector of length 2 giving the rows and columns
}
\examples{
setcolumns(x = 6)
setcolumns(x = 6, maxrow = 2)
}
\author{
Simon Frey
}
