% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothxts.r
\name{smoothdaily}
\alias{smoothdaily}
\title{Smooth a time series on daily basis}
\usage{
smoothdaily(x, upperbound = NULL)
}
\arguments{
\item{x}{An xts object.}

\item{upperbound}{(Optional) A numeric value giving the threshold below which smoothing is performed.}
}
\value{
A smoothed xts.
}
\description{
Smoothes an time series x on daily basis
}
\details{
Smoothing is performed if no value within the processed day exeeds upperbound. Smoothing is done by \eqn{(max(x)+max(y))/2}.
}
\examples{
data(runoff)
x <- smoothdaily(runoff[,1])
x <- smoothdaily(runoff[,1], upperbound = quantile(runoff[,1], 0.66))
}
\author{
Simon Frey
}
