% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothxts.r
\name{smoothXTS}
\alias{smoothXTS}
\title{Smooth a time series}
\usage{
smoothXTS(x, timesteps = 24, lowerbound = NULL, upperbound = NULL)
}
\arguments{
\item{x}{an xts object.}

\item{timesteps}{integer. Number of timestpes used for smoothing}

\item{upperbound}{numeric. (Optional). Upper boundary. Below this, smoothing will be performed.}

\item{lowerbund}{numeric. (Optional) Lower boundary. Above this smoothing will be performed.}
}
\value{
An xts object
}
\description{
Smooth a time series, optionally within two boundaries.
}
\details{
It uses the function \code{\link{embed.fullextent}} to smooth the timeseries
}
\examples{
data(runoff)
x <- runoff[,1]
summary(x)
X <- smoothXTS(x)
summary(X)
X <- smoothXTS(x, lowerbound = quantile(x,0.2), upperbound = quantile(x, 0.8))
summary(X)
}
\seealso{
\code{\link{embed.fullextent}}
}
\author{
Simon Frey
}
