% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_COSEROreg_par.r
\name{split_monthly}
\alias{split_monthly}
\title{Disaggregate monthly parameters from a COSERO parameter file}
\usage{
split_monthly(x, parameter, wnames)
}
\arguments{
\item{x}{the output from \link{read_COSERO_par}}

\item{parameter}{character. Name of the parameter that will be split and returned. Note the the parameter name must not contain numbers indicating the month. E.g. PCOR instead of PCOR1}

\item{wnames}{integer or integer vector. Position(s) of names in vector holding all COSERO Parameter names.}
}
\value{
a matrix containing 12 columns
}
\description{
Disaggregate parameters with dimensions NBIZNM to monthly columns
}
\details{
In COSERO (old) some parameters use the dimensions NBIZNN (subbasin, cell within the subbasin, month) and are stored in the parameterfile
      in one column where all NBIZ of month 1 are placed above the ones of NBIZ of month 2 and so on and so forth. In COSEROreg, however, the
      parameterfile uses distinct columns per month. This routine splits the monhtly based parameters and stores them in a matrix where every
      column represents a month.
}
\author{
Simon Frey
}
