% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splityears.r
\name{splityears}
\alias{splityears}
\title{Split an XTS object into years}
\usage{
splityears(x, type = "hydro")
}
\arguments{
\item{x}{an xts object}

\item{type}{String. Either 'hydro' for hydrological years or 'calendar' for splitting into calendar years}
}
\value{
A list where each entry contains a single year
}
\description{
Split an xts object either into hydrological or calendar years.
}
\details{
Split an xts object into years. One can chose between hydrological years starting with October 1st or calendar years starting with January 1st.
}
\examples{
 
    # load runoff data
    data(runoff)
    summary(runoff)
    
    # split data into hydrological years
    x <- splityears(runoff)
    summary(x)
    
    # split data into calendar years
    x <- splityears(runoff, "calendar")
    summary(x)
}
\seealso{
\code{\link{yearplot}}
}
\author{
Simon Frey
}
