% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{substrRight}
\alias{substrRight}
\title{Return last n characters of a string}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{character. A character string}

\item{n}{integer. Number of characters}
}
\value{
character string of the last n characters of x
}
\description{
Return last n characters of a string
}
\examples{
substrRight('hello',1)
substrRight('more than one word', 6)
}
\seealso{
\code{\link{substr}}
}
\author{
Simon Frey
}
