% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbl.r
\name{wbl}
\alias{wbl}
\title{Water balance per time span}
\usage{
wbl(sim, obs, time = "year")
}
\arguments{
\item{sim}{xts object of simulated data}

\item{obs}{xts object of observed data}

\item{time}{character string of time span.}
}
\value{
list containing xts objects of sim and obs data per time span unit
}
\description{
Calculate water balance within a time span
}
\details{
Sim and obs are subsetted into bits given by time. Accepted values are "none", "year".

   Cumulated sums are then calculated of the time spans for sim and obs and returned as a list.

   If time = year, sim and obs are subsetted in hydrological years (1st of November until 30th of October).
}
\examples{
data("area")
fpath <- system.file("extdata", "Q_output.txt", package = "TigeR")
x <- readCosero(qoutput = fpath, area = area)
sim <- x$runoff$sim.mm[,1]
obs <- x$runoff$obs.mm[,1]

wb <- wbl(sim = sim, obs = obs)
}
\seealso{
\code{\link{readCosero}}
}
\author{
Simon Frey
}
