% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.state.r
\name{write.state}
\alias{write.state}
\title{Write a COSERO state file}
\usage{
write.state(x, file = "statevar.dmp")
}
\arguments{
\item{x}{a list of length=4. The output of \code{\link{read.state}}}

\item{file}{file that will be written}
}
\description{
Write a file containing the state variables of COSERO
}
\details{
This function writes a COSERO state file to the hard disk.
}
\examples{
    x <- read.state("C:/TEMP/statevar.dmp")
    write.stat(x, file = "C:/TEMP/statevar_by_R.dmp")
}
\seealso{
\code{\link{analyse.snow.state}}

\code{\link{readstate}}
}
\author{
Simon Frey
}
