% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{write.xts}
\alias{write.xts}
\title{Write an xts object using the date format as rownames}
\usage{
write.xts(x, format = NULL, ...)
}
\arguments{
\item{x}{xts object}

\item{format}{format-style argument for formatting the date/time}

\item{...}{Additonal arguments passed to write.table}
}
\description{
Writing an xts object to a file using its date format as rownames instead of the numerical values of the date
}
\examples{
### do not run
data("runoff")
write.xts(runoff, file = tempfile())
write.xts(runoff, format = "\%d.\%m.\%Y \%H:\%M")
}
\author{
Simon Frey
}
