% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCOSEROpar.r
\name{write_COSERO_par}
\alias{write_COSERO_par}
\title{Write a COSERO parameter file}
\usage{
write_COSERO_par(x, file = "parameter_COSERO.par")
}
\arguments{
\item{x}{previously loaded parameter file}

\item{file}{character string. The filename of the new parameter file.}
}
\value{
Nothing is returned to R
}
\description{
write a previously loaded COSERO parameter file
}
\author{
Simon Frey
}
