% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_COSEROreg_par.r
\name{write_COSEROreg_par}
\alias{write_COSEROreg_par}
\title{Write a COSEROReg Parameterfile}
\usage{
write_COSEROreg_par(x, file = "Parameter_COSERO.par")
}
\arguments{
\item{x}{output from \link{read_COSERO_par}}

\item{file}{character string. Filename of the output file. If NULL the output will be returned to R.}
}
\description{
write a COSEROReg Parameterfile
}
\details{
Intended use of this function is the translation of old parameter files of COSERO into new ones used by COSEROreg
}
\examples{
    ## Do not run
    
    # Return matrix to R
    mat <- write_CoSEROreg_par(x=output_from_read_COSERO_par, file = NULL)
    
    # Write file on disk
    write_CoSEROreg_par(x=output_from_read_COSERO_par, file = "Path_to_file")
    
}
\seealso{
\link{read_COSERO_par} for reading old COSERO parameter files

\link{write_COSERO_par} for writing old COSERO parameter files
}
\author{
Simon Frey
}
