% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearplot.r
\name{yearplot}
\alias{yearplot}
\title{Plots an xts object splitted in hydrological years}
\usage{
yearplot(
  sim,
  obs,
  NB = NULL,
  plot = TRUE,
  maxrow = 4,
  dev = "dev",
  plotfile = NULL,
  ...
)
}
\arguments{
\item{sim}{xts object with simulated data from COSERO}

\item{obs}{xts object with observed data from COSERO}

\item{NB}{numeric. Subbasin to be plotted}

\item{plot}{logical. should the results be plotted or returned?}

\item{maxrow}{numeric. Maximum plots in a row}

\item{dev}{characterstring. format of the plot (valid is "dev","pdf","png")}

\item{plotfile}{characterstring. Filename of the plot if dev != "dev"}

\item{...}{additional arguments passed to plot}
}
\value{
An xts object is plotted
}
\description{
Plots an xts object splitted in hydrological years
}
\examples{
    # load runoff data
    data(runoff)
    yearplot(sim = runoff[,2], obs = runoff[,1])
}
\seealso{
\code{\link{dev.new.file}}

\code{\link{splityears}}
}
\author{
Simon Frey
}
