% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_statefiles.r
\name{zip_statefiles}
\alias{zip_statefiles}
\title{Create zipfiles from COSERO statevarfiles for use in Delft-FEWS}
\usage{
zip_statefiles(
  wd,
  selection = "all",
  prefix = "COSERO_",
  postfix = "_Update Default"
)
}
\arguments{
\item{wd}{Character string. Working directory}

\item{selection}{chracter string. Either "all" or a vector of directories within wd that will be processed.}

\item{prefix}{character string. Prefix for the name of the zipfile}

\item{postfix}{chracter string. Postfix for the name of the zipfile}
}
\value{
Zipfiles are written to the harddisk.
}
\description{
Create zipfiles from a set of statevarfiles to be used in Delft-FEWS.
}
\details{
The zipfiles are named according to the folder they are located in combined with a prefix and a postfix.
The standard filename then will be "COSERO_Alz_Update Default.zip", for instance.
}
\author{
Simon Frey
}
