% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.daily.columns.r
\name{apply.daily.columns}
\alias{apply.daily.columns}
\title{Columnwise apply a function to receive daily agreggated values}
\usage{
apply.daily.columns(x, FUN, PB = "n", ...)
}
\arguments{
\item{x}{an xts object}

\item{FUN}{an R function}

\item{PB}{a character indicating whether and what kind of progress bar should be drawn. See details.}

\item{...}{additional arguments to FUN}
}
\value{
An xts object containing daily values
}
\description{
Apply a specified function to each column of an xts object creating daily values
}
\details{
A simple mechanism to use \code{\link{apply.daily}} to each column of an xts object. 
    
    By setting PB, an optional progressbar can be drawn: "w" or "win" draws a \code{\link{winProgressBar}}, 
    "t" or "txt" draws a \code{\link{txtProgressBar}} and "n" or "none" (the default) suppresses the progressbar.
}
\author{
Simon Frey
}
