% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.daily.columns.r
\name{apply.daily.columns}
\alias{apply.daily.columns}
\title{Columnwise apply a function to receive hourly or daily agreggated values}
\usage{
apply.daily.columns(x, FUN, agg = "day", PB = "n", tz = NULL, ...)
}
\arguments{
\item{x}{an xts object}

\item{FUN}{an R function}

\item{agg}{character string 'day' or 'hour' to specifiy whether to receive daily or hourly values, respectivly}

\item{PB}{a character indicating whether and what kind of progress bar should be drawn. See details.}

\item{tz}{character specifying the time zone or NULL (the standard). If the latter, the time zone of x is used}

\item{...}{additional arguments to FUN}
}
\value{
An xts object containing daily values
}
\description{
Apply a specified function to each column of an xts object creating hourly or daily values
}
\details{
A simple mechanism to use \code{\link{apply.daily}} or \code{\link{apply.hourly}} to each column of an xts object. 
    
    By setting PB, an optional progressbar can be drawn: "w" or "win" draws a \code{\link{winProgressBar}}, 
    "t" or "txt" draws a \code{\link{txtProgressBar}} and "n" or "none" (the default) suppresses the progressbar.
}
\examples{
 
    # load precipitation input
    data(precipitation)
    head(x)
    
    # aggregate to daily values
    aday <- apply.daily.columns(x, FUN = sum, agg = 'day', PB = 'txt')
    head(aday)
}
\seealso{
\code{\link{apply.daily}}

\code{\link{apply.hourly}}
}
\author{
Simon Frey
}
