% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{apply.hourly}
\alias{apply.hourly}
\title{aggregate xts object to hourly values}
\usage{
apply.hourly(x, FUN, ...)
}
\arguments{
\item{x}{xts object}

\item{FUN}{function to apply to x}

\item{...}{additonal arguments passed to apply}
}
\description{
Aggregate an xts object to hourly values. It is a wrapper of \code{\link{period.apply}} with endpoints = "hours"
}
\seealso{
\code{\link{period.apply}}
}
\author{
Simon Frey
}
