% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{clear.duplicated}
\alias{clear.duplicated}
\title{Remove duplicated entries}
\usage{
clear.duplicated(x)
}
\arguments{
\item{x}{xts object}
}
\value{
xts object
}
\description{
Searches duplicated timestamps in an xts object. If some are found, the values of this entry and its neighbours are averaged
}
\author{
Simon Frey
}
