% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{dev.new.file}
\alias{dev.new.file}
\title{Open a new device for plotting}
\usage{
dev.new.file(device = "dev", return.device = TRUE, filename = NULL, ...)
}
\arguments{
\item{device}{any string of "dev","png","pdf", or "svg"}

\item{return.device}{logical. should the type of device be returned?}

\item{filename}{string. Where should the plot be saved? If no extention is given, it will be added according to the type of the device.}

\item{...}{additional arguments passed on to the dev fuction (e.g. \link{dev.new}, or \link{png}).}
}
\description{
Open a new device for plotting and allow some settings
}
\examples{
    d <- dev.new.file(device = "png",filename = "C:/TEMP/test")
    plot(1)
    if(d \%in\% c("png","pdf","svg")) dev.off()

}
\author{
Simon Frey
}
