% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{dev.new.file}
\alias{dev.new.file}
\title{Open new graphical device}
\usage{
dev.new.file(device = "dev", return.device = TRUE, filename = NULL, ...)
}
\arguments{
\item{device}{character string. See details.}

\item{return.device}{logical. Should the device character string be returned}

\item{filename}{character string.}

\item{...}{Additional arguments passed from other methods}
}
\value{
If return.device = TRUE, device is retured. Else nothing is returned.
}
\description{
Open a new graphical device on screen or open a pdf/png file for plotting
}
\details{
Valid arguments for device are "dev", "png" and "pdf".

    If device = "dev", the standard, then a new graphical deive is opened by calling dev.new()

    If device = "png" or "pdf" the respective file specified by filename is opend for wrting. Note that it is not automatically closed.

    Filename doesn't need to end with an ending. It will be concatenated if missing.

    Units of the devices is inches.
    A png will have the resolution of 300 dpi
}
\examples{
dev.new.file()
dev.new.file(width=7, height=2)
dev.new.file(device = "png", return.device = TRUE, filename = tempfile)
# do not run
dev.new.file(device = "pdf")

}
\author{
Simon Frey
}
