% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{missing.steps}
\alias{missing.steps}
\title{Find missing time steps in an univariate time series}
\usage{
missing.steps(x)
}
\arguments{
\item{x}{an xts object}
}
\value{
a numerical vector giving the differences of the time steps of the xts object
}
\description{
It is essentially a wrapper around \code{\link{diff.POSIXt}} returning a numeric vector. Any irregularities then can be interpreted as missing steps.
}
\author{
Simon Frey
}
