% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdiff.r
\name{peakdiff}
\alias{peakdiff}
\title{Calculate the peak difference of two hydrographs}
\usage{
peakdiff(sim, obs, timesteps, width = 12, FUN = me, nop = 5, ...)
}
\arguments{
\item{sim}{xts object. The simulated hydrograph.}

\item{obs}{xts object. The observed hydrograph.}

\item{timesteps}{numeric. How many timesteps should be considered (equal distributed around the observed peak)?}

\item{width}{numeric. Number of timesteps on eighter side of the peak that must be smaller than the peak. Corresponds to paramteter m in \code{\link{find_peaks}}}

\item{FUN}{function. Objective function to be calculated whitin the region of the peak(s).}

\item{nop}{integer. Number of peaks being evaluated.}

\item{...}{additional arguments passed to FUN}
}
\value{
a vector with the results of the objective function
}
\description{
Calculate the peak difference of two hydrographs.
}
\details{
FUN should be a function that can operate with sim and obs values and returns one single value.
    
    For detecting peaks \code{\link{find_peaks}} is used. Peaks are detected in obs only!
}
\seealso{
\code{\link{find_peaks}}
}
\author{
Simon Frey
}
