% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.state.r
\name{read.state}
\alias{read.state}
\title{Read a COSERO state file}
\usage{
read.state(file = "statevar.dmp")
}
\arguments{
\item{file}{path to a state file}
}
\value{
a list (see Details)
}
\description{
Read a file containing the state variables of COSERO
}
\details{
This function simply reads a state variable file of COSERO and returns a list containing the four parts of the file, each as a matrix:
    1) The variables on the subbasin scale (NB)
    2) The variables on the grid scale (NB, IZ)
    3) The variables on the subgrid scale, i.e. the state of the snow cover (NB,IZ,IKL)
    4) The state of the glacier runoff routing
}
\examples{
    read.state("C:/TEMP/statevar.dmp")
}
\seealso{
\code{\link{analyse.snow.state}}
}
\author{
Simon Frey
}
