% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{read.xts}
\alias{read.xts}
\title{read a textfile and save it as xts-object}
\usage{
read.xts(
  x,
  datecolumns = 1,
  format = "\%Y-\%m-\%d \%H:\%M",
  header = TRUE,
  tz = "utc",
  skip = 0,
  cut.prefix = 0,
  ...
)
}
\arguments{
\item{x}{character string. file to load}

\item{datecolumns}{integer vector. Giving the columns with date informations}

\item{format}{character string. Specify the format of the date}

\item{header}{logical. Does the data have a header?}

\item{tz}{character string. Time zone of the data}

\item{skip}{integer: the number of lines of the data file to skip before beginning to read data.}

\item{cut.prefix}{integer. Amount of leading characters that will be ignored in formatting the POSIXct object. Experimental!}

\item{...}{additional arguments from other methods passed to \code{\link{fread}}}
}
\value{
an xts object.
}
\description{
Read a text file using \code{\link{fread}} and directly save it as \code{\link{xts}} object.
}
\details{
Other than in older versions, where the standard arguments of this function were designed to read COSERO input and output files, this function by default reads tab-separated files using the time format Y-m-d H:M.
}
\examples{
fpath <- system.file("extdata","qobs_qsim.txt", package = "TigR")
out <- read.xts(x = fpath)
}
\seealso{
\code{\link{write.xts}}
}
\author{
Simon Frey
}
