% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcolstats.r
\name{rowStats}
\alias{rowStats}
\title{Calculate some simple statistic estimators on rows in a matrix or data.frame}
\usage{
rowStats(x, fun, na.rm = TRUE, q = 0.1, weights = NULL)
}
\arguments{
\item{x}{numerical matrix or data.frame.}

\item{fun}{statistical function that will be calculated. See details.}

\item{na.rm}{logical. Should NA values be removed?}

\item{q}{numerical. Quantile that will be estimated. Used only if fun == quantile.}

\item{weights}{numerical vector. Used if fun == weighted.quantile only. See details.}
}
\value{
numerical vector
}
\description{
Calculate some simple statistics like median, min, max or quantiles on rows in a matrix or data.frame object
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function uses predefined statistical function that can be used:
* min (determine the minimum value)
* max (determine the maximum value)
* mean (estimate the mean value)
* median (estimate the median value)
* quantile (estimate a certain quantile value. Note that values between 0 and 1 are accepted)
* weighted.quantile (calculation of quantiles using individual weights. \link{Quantile} from \code{DescTools} is used.)
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link{colStats}}

\code{\link{colMin}}

\code{\link{rowMin}}

\code{\link{rowMax}}

\code{\link{colMax}}
}
\author{
Simon Frey
}
