% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{unpack}
\alias{unpack}
\title{unpack an archive (zip or tar)}
\usage{
unpack(x, keeptar = TRUE, exdir = ".", ...)
}
\arguments{
\item{x}{character sting. file to be unpacked}

\item{keeptar}{logical. If TRUE only decompresses the .tar.gz archive, otherwise fully unpack it?}

\item{exdir}{character string. The directory to extract files to.}

\item{...}{arguments passed from other method}
}
\value{
The default of unzip/untar/gunzip will be returned
}
\description{
This funtcion tries to unpack a tar.gz or zip file
}
\details{
If keeptar == TRUE, the .tar.gz archive is only beeing decompressed using the \code{\link{gunzip}}-function from the R.utils-package (to the .tar-file). Note that remove is set to FALSE, unlike the default of gunzip.

   If keeptar == FALSE, the archive is fully unpacked using the \code{\link{untar}} function.
}
\seealso{
\code{\link{unzip}}, \code{\link{untar}}, \code{\link{gunzip}}
}
\author{
Simon Frey
}
