% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{unpack}
\alias{unpack}
\title{unpack a set of files}
\usage{
unpack(
  x,
  keeptar = TRUE,
  exdir = ".",
  overwrite = TRUE,
  remove = FALSE,
  cores = NULL,
  barstyle = "win",
  debug = FALSE,
  ...
)
}
\arguments{
\item{x}{vector of filenames}

\item{keeptar}{Should the tar.gz archives be extracted or extracted and untared?}

\item{exdir}{character string. The directory to extract files to.}

\item{overwrite}{logical. Should any existing files be overwritten?}

\item{remove}{logical. Should the archive be deleted after extraction? Also deletes skipped archives.}

\item{cores}{integer. Number of cores used to extract the files. Defaults to NULL which is a single core.}

\item{barstyle}{character. Style of the progressbar drawn by \code{\link{pbapply}}. Possible values are "timer", "txt", "win", "tk", or "none".}

\item{debug}{logical. Should informations about debugging be printed?}

\item{...}{arguments passed from other method}
}
\value{
The default of unzip/untar/gunzip will be returned
}
\description{
This function can handle tar.gz and zip files. One can decide if tar.gz files should be just extracted (use keeptar = TRUE) or extracted and untared (use keeptar = FALSE).
}
\seealso{
\code{\link{unzip}}, \code{\link{untar}}, \code{\link{gunzip}}
}
\author{
Simon Frey
}
